changing brainObjs to brainObj
